move code into main and separate progress dialog
